+++ packages/expo-media-library/src/next/native/NativeMediaLibraryModule.web.ts
@@ -6,6 +6,7 @@ import {
 } from 'expo';
 
 import type {
+  AlbumMetadata,
   AlbumType,
   AssetField,
   AssetFieldValueMap,
@@ -121,6 +122,9 @@ class NativeAlbumWeb implements NativeAlbumClass {
   getTitle() {
     return unavailable('Album.getTitle');
   }
+  getAssetCount() {
+    return unavailable('Album.getAssetCount');
+  }
   getType(): Promise<AlbumType> {
     return unavailable('Album.getType');
   }
@@ -152,6 +156,10 @@ class NativeAlbumWeb implements NativeAlbumClass {
 
   static getAll(): Promise<NativeAlbumClass[]> {
     return unavailable('Album.getAll');
+  }
+
+  static getAlbumsMetadata(): Promise<AlbumMetadata[]> {
+    return unavailable('Album.getAlbumsMetadata');
   }
 
   static getSmartAlbums(): Promise<NativeAlbumClass[]> {